+++ src/e3/testsuite/report/xunit.py
@@ -209,53 +209,69 @@ class XUnitImporter:
                 result.time = float(time_str) if time_str else None
                 status = TestStatus.PASS
                 message: str | None = None
+                decoding_errors: list[str] = []
 
                 # Expect at most one "error", "failure" or "skipped" element.
                 # Presence of one such element or the absence of elements
-                # enables us to determine the test status.
-                count = len(testcase)
-                assert count <= 1
-                if count == 1:
-                    status_elt = testcase[0]
-                    tag = status_elt.tag
-                    if tag == "error":
-                        status = TestStatus.ERROR
-                    elif tag == "failure":
-                        status = TestStatus.FAIL
-                    elif tag == "skipped":
-                        # xUnit reports created by py.test may contain a "type"
-                        # attribute that disambiguates between skip and xfail
-                        # results.
-                        kind = status_elt.attrib.get("type")
-                        status = (
-                            TestStatus.XFAIL
-                            if kind == "pytest.xfail"
-                            else TestStatus.SKIP
+                # enables us to determine the test status. Also allow
+                # "system-out" and "system-err" elements, for extra logging.
+                status_found = False
+                for elt in testcase:
+                    match elt.tag:
+                        case "error":
+                            status = TestStatus.ERROR
+                        case "failure":
+                            status = TestStatus.FAIL
+                        case "skipped":
+                            # xUnit reports created by py.test may contain a
+                            # "type" attribute that disambiguates between skip
+                            # and xfail results.
+                            kind = elt.attrib.get("type")
+                            status = (
+                                TestStatus.XFAIL
+                                if kind == "pytest.xfail"
+                                else TestStatus.SKIP
+                            )
+                        case "system-out" | "system-err":
+                            if isinstance(elt.text, str):
+                                result.log += f"\n\n{elt.tag}:\n"
+                                result.log += elt.text
+                            continue
+                        case tag:
+                            decoding_errors.append(f"unexpected tag: {tag}")
+                            continue
+
+                    # Execution reaches this point only for elements that
+                    # encode the test status.
+                    if status_found:
+                        decoding_errors.append("too many status elements")
+                        continue
+                    status_found = True
+                    if isinstance(elt.text, str):
+                        result.log += elt.text
+                    message = elt.attrib.get("message")
+
+                # Some XUnit producers are know to put full logs in the
+                # "message" attribute, which produces unexpected results in
+                # e3-testsuite report viewers. Keep the first line only,
+                # and cap its length if needed.
+                #
+                # If we end up modifying it, put the unmodified version in the
+                # logs, to avoid information loss.
+                if message is not None:
+                    original = message
+                    if "\n" in message:
+                        message = message.split("\n", 1)[0]
+                    if len(message) > 200:
+                        message = message[:200]
+                    if message != original:
+                        result.log.log = (
+                            "Status message was too long:\n\n"
+                            + original
+                            + "\n\n"
+                            + result.log.log
                         )
-                    else:  # all: no cover
-                        raise AssertionError(f"invalid status tag: {tag}")
-                    if isinstance(status_elt.text, str):
-                        result.log += status_elt.text
-
-                    message = status_elt.attrib.get("message")
-
-                    # Some XUnit producers are know to put full logs in the
-                    # "message" attribute, which produces unexpected results in
-                    # e3-testsuite report viewers. Keep the first line only,
-                    # and cap its length if needed.
-                    capped = False
-                    if message is not None:
-                        if "\n" in message:
-                            message = message.split("\n", 1)[0]
-                            capped = True
-                        if len(message) > 200:
-                            message = message[:200]
-                            capped = True
-                        if capped:
-                            message = message.strip() + " [...]"
-
-                else:
-                    message = None
+                        message = message.strip() + " [...]"
 
                 # Now that the "unrefined" status for this result is known,
                 # apply XFAIL, if needed.
@@ -284,6 +300,23 @@ class XUnitImporter:
                     elif status in (TestStatus.FAIL, TestStatus.ERROR):
                         status = TestStatus.XFAIL
                         message = new_message
+
+                # If we had trouble making sense of the XML entry for this test
+                # result, turn it into an error and integrate our diagnostics.
+                # This bypasses the XFAIL messages intentionally.
+                if decoding_errors:
+                    result.log += (
+                        "\n\nErrors while decoding the xUnit report for this"
+                        " testcase:\n\n"
+                    )
+                    for e in decoding_errors:
+                        result.log += f"  * {e}\n"
+                    result.log += (
+                        "\nSo turning the following into an ERROR result:\n\n"
+                        f"  {status.name}: {message}\n"
+                    )
+                    status = TestStatus.ERROR
+                    message = "xUnit report decoding error"
 
                 result.set_status(status, message)
                 self.index.save_and_add_result(result)